really getting done replacing long with unsigned int

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -10,7 +10,7 @@
 int _atoi(char *s)
 {
 	int sign = 1;
-	long result = 0;
+	unsigned int result = 0;
 
 	while (*s)
 	{
